Update Adafruit_ADS1015.cpp
Line 67 did not make sense. The higher byte was shifted to the lower byte and than ored to the lower byte.
What needs to happen is, the lower byte and higher byte of the config word get swapped.
The change was tested with hardware, the previous version didnt work, the new version now works.